pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sigmark"
version = "0.3.0"
description = "GPG signing for static site markdown content"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Alex Towell", email = "lex@metafunctor.com" }
]
keywords = ["gpg", "markdown", "static-site", "hugo", "signing", "pgp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
]

dependencies = [
    "click>=8.0",
    "ruamel.yaml>=0.17",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
sigmark = "sigmark.cli:main"

[project.urls]
Repository = "https://github.com/queelius/sigmark"
Issues = "https://github.com/queelius/sigmark/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/sigmark"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build]
include = ["src/sigmark/**/*.py"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/sigmark"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
